# docs/public-api/references/swagger/process/flowable-swagger-process.yaml
...
swagger: "2.0"
info:
  description: "# flowable / flow??b(??)l /\n\n- a compact and highly efficient workflow\
    \ and Business Process Management (BPM) platform for developers, system admins\
    \ and business users.\n- a lightning fast, tried and tested BPMN 2 process engine\
    \ written in Java. It is Apache 2.0 licensed open source, with a committed community.\n\
    - can run embedded in a Java application, or as a service on a server, a cluster,\
    \ and in the cloud. It integrates perfectly with Spring. With a rich Java and\
    \ REST API, it is the ideal engine for orchestrating human or system activities."
  version: "v1"
  title: "Flowable Process REST API"
  contact:
    name: "Flowable"
    url: "http://www.flowable.org/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/flowable-rest/service"
...
schemes:
- "http"
- "https"
paths:
...
  /identity/groups:
    get:
      tags:
      - "Groups"
      summary: "List groups"
      description: ""
      operationId: "listGroups"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "Only return group with the given id"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "Only return groups with the given name"
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "Only return groups with the given type"
        required: false
        type: "string"
      - name: "nameLike"
        in: "query"
        description: "Only return groups with a name like the given value. Use % as\
          \ wildcard-character."
        required: false
        type: "string"
      - name: "member"
        in: "query"
        description: "Only return groups which have a member with the given username."
        required: false
        type: "string"
      - name: "potentialStarter"
        in: "query"
        description: "Only return groups which members are potential starters for\
          \ a process-definition with the given id."
        required: false
        type: "string"
      - name: "sort"
        in: "query"
        description: "Property to sort on, to be used together with the order."
        required: false
        type: "string"
        enum:
        - "id"
        - "name"
        - "type"
      responses:
        200:
          description: "Indicates the requested groups were returned."
          schema:
            $ref: "#/definitions/DataResponseGroupResponse"
      security:
      - basicAuth: []
    post:
      tags:
      - "Groups"
      summary: "Create a group"
      description: ""
      operationId: "createGroup"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/GroupRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GroupResponse"
        201:
          description: "Indicates the group was created."
        400:
          description: "Indicates the id of the group was missing."
      security:
      - basicAuth: []
  /identity/groups/{groupId}:
    get:
      tags:
      - "Groups"
      summary: "Get a single group"
      description: ""
      operationId: "getGroup"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Indicates the group exists and is returned."
          schema:
            $ref: "#/definitions/GroupResponse"
        404:
          description: "Indicates the requested group does not exist."
      security:
      - basicAuth: []
    put:
      tags:
      - "Groups"
      summary: "Update a group"
      description: "All request values are optional. For example, you can only include\
        \ the name attribute in the request body JSON-object, only updating the name\
        \ of the group, leaving all other fields unaffected. When an attribute is\
        \ explicitly included and is set to null, the group-value will be updated\
        \ to null."
      operationId: "updateGroup"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/GroupRequest"
      responses:
        200:
          description: "Indicates the group was updated."
          schema:
            $ref: "#/definitions/GroupResponse"
        404:
          description: "Indicates the requested group was not found."
        409:
          description: "Indicates the requested group was updated simultaneously."
      security:
      - basicAuth: []
    delete:
      tags:
      - "Groups"
      summary: "Delete a group"
      description: ""
      operationId: "deleteGroup"
      parameters:
      - name: "groupId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Indicates the group was found and  has been deleted. Response-body\
            \ is intentionally empty."
        404:
          description: "Indicates the requested group does not exist."
      security:
      - basicAuth: []
  /identity/groups/{groupId}/members:
    post:
      tags:
      - "Groups"
      summary: "Add a member to a group"
      description: ""
      operationId: "createMembership"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/MembershipRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/MembershipResponse"
        201:
          description: "Indicates the group was found and the member has been added."
        400:
          description: "Indicates the userId was not included in the request body."
        404:
          description: "Indicates the requested group was not found."
        409:
          description: "Indicates the requested user is already a member of the group."
      security:
      - basicAuth: []
  /identity/groups/{groupId}/members/{userId}:
    delete:
      tags:
      - "Groups"
      summary: "Delete a member from a group"
      description: ""
      operationId: "deleteMembership"
      parameters:
      - name: "groupId"
        in: "path"
        required: true
        type: "string"
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Indicates the group was found and the member has been deleted.\
            \ The response body is left empty intentionally."
        404:
          description: "Indicates the requested group was not found or that the user\
            \ is not a member of the group. The status description contains additional\
            \ information about the error."
      security:
      - basicAuth: []
...
  /identity/users/{userId}/picture:
    get:
      tags:
      - "Users"
      summary: "Get a user's picture"
      description: "The response body contains the raw picture data, representing\
        \ the user's picture. The Content-type of the response corresponds to the\
        \ mimeType that was set when creating the picture."
      operationId: "getUserPicture"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Indicates the user was found and has a picture, which is returned\
            \ in the body."
          schema:
            type: "array"
            items:
              type: "string"
              format: "byte"
        404:
          description: "Indicates the requested user was not found or the user does\
            \ not have a profile picture. Status-description contains additional information\
            \ about the error."
      security:
      - basicAuth: []
    put:
      tags:
      - "Users"
      summary: "Updating a user's picture"
      description: "The request should be of type multipart/form-data. There should\
        \ be a single file-part included with the binary value of the picture. On\
        \ top of that, the following additional form-fields can be present:\n\nmimeType:\
        \ Optional mime-type for the uploaded picture. If omitted, the default of\
        \ image/jpeg is used as a mime-type for the picture."
      operationId: "updateUserPicture"
      consumes:
      - "multipart/form-data"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      - name: "file"
        in: "formData"
        description: "Picture to update"
        required: true
        type: "file"
      responses:
        204:
          description: "Indicates the user was found and the picture has been updated.\
            \ The response-body is left empty intentionally."
        404:
          description: "Indicates the requested user was not found."
      security:
      - basicAuth: []
...
  /repository/deployments/{deploymentId}/resources:
    get:
      tags:
      - "Deployment"
      summary: "List resources in a deployment"
      description: "The dataUrl property in the resulting JSON for a single resource\
        \ contains the actual URL to use for retrieving the binary resource."
      operationId: "listDeploymentResources"
      produces:
      - "application/json"
      parameters:
      - name: "deploymentId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Indicates the deployment was found and the resource list has\
            \ been returned."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DeploymentResourceResponse"
        404:
          description: "Indicates the requested deployment was not found."
      security:
      - basicAuth: []
...
  /repository/models/{modelId}/source:
    get:
      tags:
      - "Models"
      summary: "Get the editor source for a model"
      description: "Response body contains the model?s raw editor source. The response?\
        s content-type is set to application/octet-stream, regardless of the content\
        \ of the source."
      operationId: "getModelBytes"
      parameters:
      - name: "modelId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Indicates the model was found and source is returned."
          schema:
            type: "array"
            items:
              type: "string"
              format: "byte"
        404:
          description: "Indicates the requested model was not found."
      security:
      - basicAuth: []
    put:
      tags:
      - "Models"
      summary: "Set the editor source for a model"
      description: "Response body contains the model?s raw editor source. The response?\
        s content-type is set to application/octet-stream, regardless of the content\
        \ of the source."
      operationId: "setModelSource"
      consumes:
      - "multipart/form-data"
      parameters:
      - name: "modelId"
        in: "path"
        required: true
        type: "string"
      - name: "file"
        in: "formData"
        required: true
        type: "file"
      responses:
        204:
          description: "Indicates the model was found and the source has been updated."
        404:
          description: "Indicates the requested model was not found."
      security:
      - basicAuth: []
...
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  DataResponseGroupResponse:
    type: "object"
    properties:
      data:
        type: "array"
        items:
          $ref: "#/definitions/GroupResponse"
      total:
        type: "integer"
        format: "int64"
      start:
        type: "integer"
        format: "int32"
      sort:
        type: "string"
      order:
        type: "string"
      size:
        type: "integer"
        format: "int32"
...
  DeploymentResourceResponse:
    type: "object"
    properties:
      id:
        type: "string"
        example: "diagrams/my-process.bpmn20.xml"
      url:
        type: "string"
        example: "http://localhost:8081/flowable-rest/service/repository/deployments/10/resources/diagrams%2Fmy-process.bpmn20.xml"
        description: "For a single resource contains the actual URL to use for retrieving\
          \ the binary resource"
      contentUrl:
        type: "string"
        example: "http://localhost:8081/flowable-rest/service/repository/deployments/10/resourcedata/diagrams%2Fmy-process.bpmn20.xml"
      mediaType:
        type: "string"
        example: "text/xml"
        description: "Contains the media-type the resource has. This is resolved using\
          \ a (pluggable) MediaTypeResolver and contains, by default, a limited number\
          \ of mime-type mappings."
      type:
        type: "string"
        example: "processDefinition"
        description: "Type of resource"
        enum:
        - "resource"
        - "processDefinition"
        - "processImage"
...
  GroupRequest:
    type: "object"
    properties:
      id:
        type: "string"
        example: "testgroup"
      url:
        type: "string"
        example: "http://localhost:8182/identity/groups/testgroup"
      name:
        type: "string"
        example: "Test group"
      type:
        type: "string"
        example: "Test type"
      nameChanged:
        type: "boolean"
      typeChanged:
        type: "boolean"
  GroupResponse:
    type: "object"
    properties:
      id:
        type: "string"
        example: "testgroup"
      url:
        type: "string"
        example: "http://localhost:8182/identity/groups/testgroup"
      name:
        type: "string"
        example: "Test group"
      type:
        type: "string"
        example: "Test type"
...
  MembershipRequest:
    type: "object"
    properties:
      userId:
        type: "string"
        example: "kermit"
  MembershipResponse:
    type: "object"
    properties:
      userId:
        type: "string"
        example: "kermit"
      url:
        type: "string"
        example: "http://localhost:8182/identity/groups/sales/members/kermit"
      groupId:
        type: "string"
        example: "sales"
...